Move Software Model Check task to base plugin
Previously, a check task was only added if the testing base was added, causing confusion to users who wanted to add their own check tasks which are not directly unit tests.
